Move the Enterprise CR getters into pkg/enterprise/utils
The shared controller utils package no longer reads Enterprise-only CRs, and core code importing anything under pkg/enterprise is now a boundary violation too.
