+++ pkg/controller/logcollector/logcollector_controller.go
@@ -42,6 +42,7 @@ import (
 	"github.com/tigera/operator/pkg/controller/utils"
 	"github.com/tigera/operator/pkg/controller/utils/imageset"
 	"github.com/tigera/operator/pkg/ctrlruntime"
+	eutils "github.com/tigera/operator/pkg/enterprise/utils"
 	"github.com/tigera/operator/pkg/render"
 	rcertificatemanagement "github.com/tigera/operator/pkg/render/certificatemanagement"
 	rmeta "github.com/tigera/operator/pkg/render/common/meta"
@@ -399,7 +400,7 @@ func (r *ReconcileLogCollector) Reconcile(ctx context.Context, request reconcile
 	}
 	managedCluster := managementClusterConnection != nil
 
-	managementCluster, err := utils.GetManagementCluster(ctx, r.client)
+	managementCluster, err := eutils.GetManagementCluster(ctx, r.client)
 	if err != nil {
 		r.status.SetDegraded(operatorv1.ResourceReadError, "Error reading ManagementCluster", err, reqLogger)
 		return reconcile.Result{}, err
@@ -455,7 +456,7 @@ func (r *ReconcileLogCollector) Reconcile(ctx context.Context, request reconcile
 		linseedCertNamespace = instance.Spec.MultiTenantManagementClusterNamespace
 
 		// Make sure that a tenant actually exists in the configured namespace before continuing.
-		tenant, _, err = utils.GetTenant(ctx, r.opts.MultiTenant, r.client, instance.Spec.MultiTenantManagementClusterNamespace)
+		tenant, _, err = eutils.GetTenant(ctx, r.opts.MultiTenant, r.client, instance.Spec.MultiTenantManagementClusterNamespace)
 		if err != nil {
 			r.status.SetDegraded(operatorv1.ResourceNotReady, fmt.Sprintf("Failed to retrieve tenant in ns %s", instance.Spec.MultiTenantManagementClusterNamespace), err, reqLogger)
 			return reconcile.Result{}, err
@@ -632,7 +633,7 @@ func (r *ReconcileLogCollector) Reconcile(ctx context.Context, request reconcile
 	}
 
 	// Check if non-cluster host feature is enabled.
-	nonclusterhost, err := utils.GetNonClusterHost(ctx, r.client)
+	nonclusterhost, err := eutils.GetNonClusterHost(ctx, r.client)
 	if err != nil {
 		r.status.SetDegraded(operatorv1.ResourceReadError, "Failed to query NonClusterHost resource", err, reqLogger)
 		return reconcile.Result{}, err
