+++ pkg/enterprise/apiserver/extension.go
@@ -45,6 +45,7 @@ import (
 	"github.com/tigera/operator/pkg/ctrlruntime"
 	"github.com/tigera/operator/pkg/dns"
 	eoptions "github.com/tigera/operator/pkg/enterprise/options"
+	eutils "github.com/tigera/operator/pkg/enterprise/utils"
 	"github.com/tigera/operator/pkg/extensions"
 	"github.com/tigera/operator/pkg/render"
 	"github.com/tigera/operator/pkg/render/common/authentication"
@@ -195,12 +196,12 @@ func (e *Extension) ExtendInputs(ctx context.Context, ci controller.Inputs) (con
 		return ci, nil, extensions.Degradedf(operatorv1.ResourceCreateError, "unable to create the trusted bundle: %w", err)
 	}
 
-	applicationLayer, err := utils.GetApplicationLayer(ctx, ci.Client)
+	applicationLayer, err := eutils.GetApplicationLayer(ctx, ci.Client)
 	if err != nil {
 		return ci, nil, extensions.Degradedf(operatorv1.ResourceReadError, "error reading ApplicationLayer: %w", err)
 	}
 
-	managementCluster, err := utils.GetManagementCluster(ctx, ci.Client)
+	managementCluster, err := eutils.GetManagementCluster(ctx, ci.Client)
 	if err != nil {
 		return ci, nil, extensions.Degradedf(operatorv1.ResourceReadError, "error reading ManagementCluster: %w", err)
 	}
@@ -249,12 +250,12 @@ func (e *Extension) ExtendInputs(ctx context.Context, ci controller.Inputs) (con
 	// Authentication: when a Dex-backed Authentication CR is ready, add its cert to the
 	// bundle and build the key validator config for the query server and the policy.
 	var keyValidatorConfig authentication.KeyValidatorConfig
-	authenticationCR, err := utils.GetAuthentication(ctx, ci.Client)
+	authenticationCR, err := eutils.GetAuthentication(ctx, ci.Client)
 	if err != nil && !apierrors.IsNotFound(err) {
 		return ci, nil, extensions.Degradedf(operatorv1.ResourceReadError, "error while fetching Authentication: %w", err)
 	}
 	if authenticationCR != nil && authenticationCR.Status.State == operatorv1.TigeraStatusReady {
-		if utils.DexEnabled(authenticationCR) {
+		if eutils.DexEnabled(authenticationCR) {
 			certificate, err := ci.CertificateManager.GetCertificate(ci.Client, render.DexTLSSecretName, common.OperatorNamespace())
 			if err != nil {
 				return ci, nil, extensions.Degradedf(operatorv1.CertificateError, "failed to retrieve %s: %w", render.DexTLSSecretName, err)
@@ -263,7 +264,7 @@ func (e *Extension) ExtendInputs(ctx context.Context, ci controller.Inputs) (con
 			}
 			trustedBundle.AddCertificates(certificate)
 		}
-		keyValidatorConfig, err = utils.GetKeyValidatorConfig(ctx, ci.Client, authenticationCR, ci.RenderInputs.ClusterDomain, false)
+		keyValidatorConfig, err = eutils.GetKeyValidatorConfig(ctx, ci.Client, authenticationCR, ci.RenderInputs.ClusterDomain, false)
 		if err != nil {
 			return ci, nil, extensions.Degradedf(operatorv1.ResourceReadError, "failed to get KeyValidator config: %w", err)
 		}
@@ -316,7 +317,7 @@ func (e *Extension) ExtendInputs(ctx context.Context, ci controller.Inputs) (con
 	// is covered by its own binding.
 	var bindingNamespaces []string
 	if e.opts.MultiTenant {
-		bindingNamespaces, err = utils.TenantNamespaces(ctx, ci.Client, nil)
+		bindingNamespaces, err = eutils.TenantNamespaces(ctx, ci.Client, nil)
 		if err != nil {
 			return ci, nil, extensions.Degradedf(operatorv1.ResourceReadError, "error reading tenant namespaces: %w", err)
 		}
