+++ pkg/enterprise/utils/auth.go
