+++ pkg/enterprise/utils/auth_cloud.go
